refactor: 배틀 신청 목록 이동 버튼, UI/UX 개선
