+++ apps/mac/src/renderer/src/features/competition/ui/rival-competition/battle/Battle.style.ts
@@ -6,6 +6,7 @@ import Date from "@/features/home/assets/home/date.svg";
 import BackArrow from "@/features/home/assets/home/back.svg";
 import Fire from "@/features/home/assets/home/fire.svg";
 import Checked from "@/features/home/assets/home/check-box.svg";
+import DetailArrow from "@/shared/ui/assets/front.svg";
 
 const flexRow = css`
   display: flex;
@@ -567,4 +568,19 @@ export const ErrorText = styled.p`
   ${font.label.medium};
   color: ${palette.red[60]};
   width: 100%;
+`;
+
+export const DetailArrowIcon = styled(DetailArrow)`
+  width: 0.5rem;
+  height: 1rem;
+`;
+
+export const ArrowBox = styled.button`
+  ${font.label.medium}
+  color: ${({ theme }) => theme.label.alternative};
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 0.25rem;
+  cursor: pointer;
 `;
