+++ apps/mac/src/renderer/src/features/home/ui/transition/Transition.style.ts
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 import { font } from "@clash/design-tokens/font";
-import DetailArrow from "../../assets/home/front.svg";
+import DetailArrow from "../../../../shared/ui/assets/front.svg";
 
 const MAX_BAR_HEIGHT = "clamp(5rem, 20vw, 8rem)";
